Repository Insight, Service Insight

+++ src/main/java/tc3005b224/amazonconnectinsights/repository/InsightsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import tc3005b224.amazonconnectinsights.models_sql.Insights;
 
 @Repository
-public interface InsightsRepository extends CrudRepository<Insights, Integer> {
+public interface InsightsRepository extends CrudRepository<Insights, Short> {
 
 }
